+++ packages/ice/src/constant.ts
@@ -37,3 +37,34 @@ export const TARGETS = [
   WEEX,
   ...MINIAPP_TARGETS,
 ];
+
+export const RUNTIME_EXPORTS = [
+  {
+    specifier: ['Link', 'Outlet', 'useParams', 'useSearchParams', 'useLocation', 'useNavigate'],
+    source: '@ice/runtime/router',
+  },
+  {
+    specifier: [
+      'defineAppConfig',
+      'useAppData',
+      'useData',
+      'useConfig',
+      'Meta',
+      'Title',
+      'Links',
+      'Scripts',
+      'Data',
+      'Main',
+      'history',
+      'KeepAliveOutlet',
+      'useMounted',
+      'ClientOnly',
+      'withSuspense',
+      'useSuspenseData',
+      'defineDataLoader',
+      'defineServerDataLoader',
+      'defineStaticDataLoader',
+    ],
+    source: '@ice/runtime',
+  },
+];